test: add test_send_transaction_exceeded_maximum_ancestors_count unit test
