+++ rpc/src/tests/mod.rs
@@ -1,20 +1,28 @@
-use crate::RpcServer;
-use ckb_chain::chain::ChainController;
+use crate::{RpcServer, ServiceBuilder};
+use ckb_app_config::{BlockAssemblerConfig, NetworkConfig, RpcConfig, RpcModule};
+use ckb_chain::chain::{ChainController, ChainService};
 use ckb_dao::DaoCalculator;
+use ckb_jsonrpc_types::ScriptHashType;
+use ckb_launcher::SharedBuilder;
+use ckb_network::{DefaultExitHandler, NetworkService, NetworkState};
 use ckb_shared::{Shared, Snapshot};
 use ckb_store::ChainStore;
-use ckb_test_chain_utils::{always_success_cell, always_success_cellbase};
+use ckb_test_chain_utils::{
+    always_success_cell, always_success_cellbase, always_success_consensus,
+};
 use ckb_types::{
     core::{
-        cell::resolve_transaction, BlockBuilder, BlockView, HeaderView, TransactionBuilder,
-        TransactionView,
+        cell::resolve_transaction, BlockBuilder, BlockView, FeeRate, HeaderView,
+        TransactionBuilder, TransactionView,
     },
+    h256,
     packed::{CellInput, OutPoint},
     prelude::*,
 };
+use reqwest::blocking::Client;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use std::{collections::HashSet, fmt};
+use std::{collections::HashSet, fmt, sync::Arc};
 
 mod error;
 mod examples;
@@ -125,4 +133,110 @@ fn always_success_transaction() -> TransactionView {
         .output_data(always_success_cell_data.to_owned().pack())
         .witness(always_success_script.clone().into_witness())
         .build()
+}
+
+// setup a chain with 20 blocks and enable `Chain`, `Miner` and `Pool` rpc modules for unit test
+// there is a similar fn `setup_rpc_test_suite` which enables all rpc modules, may be refactored into one fn with different paramsters in other PRs
+fn setup() -> RpcTestSuite {
+    let (shared, mut pack) = SharedBuilder::with_temp_db()
+        .consensus(always_success_consensus())
+        .block_assembler_config(Some(BlockAssemblerConfig {
+            code_hash: h256!("0x0"),
+            args: Default::default(),
+            hash_type: ScriptHashType::Data,
+            message: Default::default(),
+            use_binary_version_as_message_prefix: false,
+            binary_version: "TEST".to_string(),
+        }))
+        .build()
+        .unwrap();
+    let chain_controller =
+        ChainService::new(shared.clone(), pack.take_proposal_table()).start::<&str>(None);
+
+    // Start network services
+    let dir = tempfile::tempdir()
+        .expect("create tempdir failed")
+        .path()
+        .to_path_buf();
+    let network_controller = {
+        let network_config = NetworkConfig {
+            path: dir,
+            ping_interval_secs: 1,
+            ping_timeout_secs: 1,
+            connect_outbound_interval_secs: 1,
+            ..Default::default()
+        };
+        let network_state =
+            Arc::new(NetworkState::from_config(network_config).expect("Init network state failed"));
+        NetworkService::new(
+            Arc::clone(&network_state),
+            Vec::new(),
+            Vec::new(),
+            shared.consensus().identify_name(),
+            "0.1.0".to_string(),
+            DefaultExitHandler::default(),
+        )
+        .start(shared.async_handle())
+        .expect("Start network service failed")
+    };
+
+    pack.take_tx_pool_builder()
+        .start(network_controller.clone());
+
+    // Build chain, insert 20 blocks
+    let mut parent = shared.consensus().genesis_block().clone();
+
+    for _ in 0..20 {
+        let block = next_block(&shared, &parent.header());
+        chain_controller
+            .process_block(Arc::new(block.clone()))
+            .expect("processing new block should be ok");
+        parent = block;
+    }
+
+    // Start rpc services
+    let rpc_config = RpcConfig {
+        listen_address: "127.0.0.1:0".to_owned(),
+        tcp_listen_address: None,
+        ws_listen_address: None,
+        max_request_body_size: 20_000_000,
+        threads: None,
+        modules: vec![
+            RpcModule::Chain,
+            RpcModule::Miner,
+            RpcModule::Pool,
+            RpcModule::IntegrationTest,
+        ],
+        reject_ill_transactions: false,
+        enable_deprecated_rpc: false,
+        extra_well_known_lock_scripts: vec![],
+        extra_well_known_type_scripts: vec![],
+    };
+
+    let builder = ServiceBuilder::new(&rpc_config)
+        .enable_chain(shared.clone())
+        .enable_pool(shared.clone(), FeeRate::zero(), true, vec![], vec![])
+        .enable_miner(
+            shared.clone(),
+            network_controller.clone(),
+            chain_controller.clone(),
+            true,
+        )
+        .enable_integration_test(shared.clone(), network_controller, chain_controller.clone());
+    let io_handler = builder.build();
+
+    let rpc_server = RpcServer::new(rpc_config, io_handler, shared.notify_controller());
+    let rpc_uri = format!(
+        "http://{}:{}/",
+        rpc_server.http_address().ip(),
+        rpc_server.http_address().port()
+    );
+    let rpc_client = Client::new();
+    RpcTestSuite {
+        shared,
+        chain_controller,
+        rpc_server,
+        rpc_uri,
+        rpc_client,
+    }
 }
