+++ rpc/src/tests/module/pool.rs
@@ -1,7 +1,19 @@
-use ckb_test_chain_utils::ckb_testnet_consensus;
-use ckb_types::{core, packed, prelude::*};
+use std::{thread::sleep, time::Duration};
 
-use crate::module::pool::WellKnownScriptsOnlyValidator;
+use ckb_store::ChainStore;
+use ckb_test_chain_utils::always_success_cell;
+use ckb_test_chain_utils::ckb_testnet_consensus;
+use ckb_types::{
+    core::{self, Capacity, TransactionBuilder},
+    packed::{self, CellDep, CellInput, CellOutputBuilder, OutPoint},
+    prelude::*,
+};
+use serde_json::json;
+
+use crate::{
+    module::pool::WellKnownScriptsOnlyValidator,
+    tests::{always_success_transaction, setup, RpcTestRequest},
+};
 
 #[test]
 fn test_default_outputs_validator() {
@@ -122,6 +134,67 @@ fn test_default_outputs_validator() {
         );
         assert!(validator.validate(&tx).is_err());
     }
+}
+
+#[test]
+#[ignore]
+fn test_send_transaction_exceeded_maximum_ancestors_count() {
+    let suite = setup();
+
+    let store = suite.shared.store();
+    let tip = store.get_tip_header().unwrap();
+    let tip_block = store.get_block(&tip.hash()).unwrap();
+    let mut parent_tx_hash = tip_block.transactions().get(0).unwrap().hash();
+
+    // generate 30 child-spends-parent txs
+    for i in 0..30 {
+        let input = CellInput::new(OutPoint::new(parent_tx_hash.clone(), 0), 0);
+        let output = CellOutputBuilder::default()
+            .capacity(Capacity::bytes(1000 - i).unwrap().pack())
+            .lock(always_success_cell().2.clone())
+            .build();
+        let cell_dep = CellDep::new_builder()
+            .out_point(OutPoint::new(always_success_transaction().hash(), 0))
+            .build();
+        let tx = TransactionBuilder::default()
+            .input(input)
+            .output(output)
+            .output_data(Default::default())
+            .cell_dep(cell_dep)
+            .build();
+        let new_tx: ckb_jsonrpc_types::Transaction = tx.data().into();
+        suite.rpc(&RpcTestRequest {
+            id: 42,
+            jsonrpc: "2.0".to_string(),
+            method: "send_transaction".to_string(),
+            params: vec![json!(new_tx), json!("passthrough")],
+        });
+        parent_tx_hash = tx.hash();
+    }
+
+    // 30 txs will be added to proposal list, block template updating is async, 1 second sleeping is enough even on poor CI server
+    for _ in 0..2 {
+        sleep(Duration::from_secs(1));
+        suite.rpc(&RpcTestRequest {
+            id: 42,
+            jsonrpc: "2.0".to_string(),
+            method: "generate_block".to_string(),
+            params: vec![],
+        });
+    }
+
+    // the default value of pool config `max_ancestors_count` is 25, only 25 txs will be added to committed list of the block template
+    sleep(Duration::from_secs(1));
+    let response = suite.rpc(&RpcTestRequest {
+        id: 42,
+        jsonrpc: "2.0".to_string(),
+        method: "get_block_template".to_string(),
+        params: vec![],
+    });
+    assert_eq!(
+        25,
+        response.result["transactions"].as_array().unwrap().len()
+    );
 }
 
 fn build_tx(
